Create countriesSouthAmerica.json
This .json file combines all the data from individual country .geo.json files (i.e., arg.geo.json). It includes the "cca2" property, which was there in the original files (i.e., "arg" for argentina). I have also added a "name" property, which gives the common name for the country, with the first letter capitalized. Note that more properties can be added to each country later on, if needed.